requirements.txt
Django~=3.2.4
django-admin-sortable2~=1.0
django-tinymce~=3.3.0
requests~=2.25.1